Switch to defmt and probe-runner
Although this doesn't yet work with gdb (cargo-embed proports to do this), it
does let us simply run programs and includes a lightweight logging mechanism.

`cargo run` will run on the target. `DEFMT_LOG` can be set in the environment
when compiling to set the highest log level that will be included in the code.