Create a main for the STM32F407VGTx disco board
Bring in the embassy project.  We probably don't actually need the git version
of embassy, so TODO: Use released embassy.  Convert the benchmark framework to
use the embassy time.  This makes the code too large for a bootloader, but
should be fine for things like benchmarking and experiements.

This does require a nightly version of rustc, due to the embassy main macro.  We
also take advantage of being able to use async closures on nightly.  Presumably,
these features will be stabilized at some point.