Store `flash` directly inside of Tlv struct
Instead of passing it through, and hoping the user passes the correct flash
device, store a RefCell of the flash storage device. This will also prevent the
outside code from manipulating the flash during the Tlv operations.  There is a
new method `.into_storage()` to extract this when finished using the Tlv, for
when it is needed again.